Moved http2 files to http2 directory
